[TH-513] 상점상세 리뷰 및 액션 구현
- 리뷰리스트 구현
- 신고 구현
- 좋아요 구현

+++ core/network/src/main/java/com/threedollar/network/data/store/NewsPost.kt
@@ -33,7 +33,7 @@ data class Section(
     val url: String,
 
     @SerializedName("ratio")
-    val ratio: Int
+    val ratio: Float
 )
 
 data class Sticker(
