+++ home/domain/src/main/java/com/home/domain/data/store/NewsPostModel.kt
@@ -17,5 +17,5 @@ enum class SectionTypeModel {
 data class SectionModel(
     val sectionType: SectionTypeModel,
     val url: String,
-    val ratio: Int
+    val ratio: Float
 )
